+++ src/qazmorph/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "Morpheme",
     "Token",
 ]
-__version__ = "0.2.1"
+__version__ = "0.2.2"